first commit
this repo used to be called Misc_tools, but there was some sensitive
data in the repo and we were having a hard time removing it :/

so, i cloned Misc_tools repo locally, then removed the repo from
Github, then created this repo (tools), and copied all of the code from
Misc_tools over to tools (minus the stuff we didn’t want to include.

should be good now!