cleaning up reprint_labels.sh
some of this is just housekeeping, but i also got rid of the queue
checking stuff in reprint_labels.sh. it seems like it’ll be me doing
this for the foreseeable future, and anyway we’re not doing queue
checking on print_labels.sh, so we’ll just need to keep an eye on the
printed BUT unboxed radios periodically to make sure it’s not out of
hand.

one not-bad idea would be to set up a cron task on my Pi that checks to
see how many labels fit that description every night. in theory that
should always be zero, but in practice i bet it’ll be 1 a fair amount
of the time due to users not opting for the “restart without printing”
option at the end of the tulip process. we’d have to watch out for that.