starting to professionalize import_orders_from_csv.rb
@gabe-ochoa hey can you check this out for me? my thoughts are:

- read in `source` and `auth_token` as arguments
- for the radio stations i can just hack the CSV by bringing it into
google docs and renaming the headers to match our default format (which
can be what is in the script currently)
- the real task is to keep this script working for Kickstarter while we
also make it functional for Uncommon Goods

TODO:
- first, check what i’m doing here - i have a suspicion it’s not
functional as is (i’m on a plane and can’t actually run it now)
- add a call to perso.py (or something similar) to confirm that the
shipping country and the frequency are kosher. if they don’t align,
raise a flag and don’t import the order, so that the label isn’t then
printed and we get a failure on the factory floor.

^ that second point is a really-nice-to-have, given what i saw this
week (labels are printed but can’t be processed)