started sketching out the path forward
getopt does not seem to do what we need, and argparse looks more
comprehensive. switching to that.

making default_firmware and default_eeprom a bit easier to handle. this
will allow the user to specify a full path to firmware and/or eeprom
files if they want to override.

sketched out how our conditionals should work to give the exit codes we
need.